+++ ChangeLog.md
@@ -9,6 +9,7 @@
 - 'install': experimental support for rpm-ostree overlaying (--rpm-ostree)
 - 'install': allow noarch tasks
 - 'install': allow package filtering in --list mode
+- 'install': --skip-existing to leave already installed subpackages untouched
 - 'progress' records the largest finished build.log size
 - 'progress': exclude srpm from max log sizes
 - 'progress': show pkg name instead of "unknown" nvr early on